Individualized Properties, Part 1
Hey. Ever want to set a property of an object relative to other
properties of itself or its parents? See normally that’d be a huge
hassle. Well, it is a huge hassle, the difference is that now Fashion
is handling the hassle for you.

This commit adds detection, basic parsing and CSS actualization for
individualized properties.

Also added unit tests for the CSS actualizer, to be safe, and fixed a
bug where it wouldn’t handle multipart properties right (those things
suck)