The pin property
The pin property allows you to position an object within its parent
(including horizontal and vertical centering).

Current limitations: It does not refresh when the object or its parent
object change size. It requires the parent object to be position:
relative already, and will always make the selected object position:
absolute.

So it’s got some problems but, hey, it works!