Improvements to triggers
Triggers are now split into 2 categories, one that captures the event
and one that doesn’t. A new runtime mode on triggers ensures that
changing variables won’t cause the property to be re-called (which may
even obsolete the need for the attribute, I’ll need to test it more).
Also fixed the compiler’s ability to distinguish the non minified
fashion.live.js file.