Direct Property Binding - Actualizer
The actualizer used to bind variables and globals to selectors, meaning
the entire selector would have to get recreated when the variable
changed. Now, variables and globals are bound to specific properties of
specific selectors, which involves substantial changes throughout the
entire system. The runtime still needs to get updated to support this
system.