Per-scope binding
The variable binding system has been tweaked to carry the particular
scope each selector is bound to. Changing a variable on the top level
now only changes selectors bound to the top level of that variable,
selectors within scopes will not be affected. At some point in the
future there may be a function developers can call to set all scopes of
a variable to a specific value.