Function Watchers & Refactored O.o()
Returning from a 2 week hiatus with a big update! Functions can now
have watchers in much the same way as globals. Function watchers are
dependent on the function’s arguments, and are automatically
regenerated when an argument’s value changes. This is demonstrated
using 3 new color functions. The DOM watchers has been rewritten to not
include hard links to other runtime libraries, and to include a very
simple polyfill for older browsers. Removed the Live runtime mode as it
is no longer necessary with function watchers. Moved some redundant
code from variables and globals into the selectors library, and
rearranged that a bit to be clearer.